requirements.txt
asgiref==3.5.2
attrs==22.1.0
az-iranian-bank-gateways==1.6.18
cached-property==1.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
cryptography==37.0.4
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.0.0
dj-email-url==1.0.5
Django==4.1
django-allauth==0.51.0
django-cache-url==3.4.2
django-ckeditor==6.5.1
django-crispy-forms==1.14.0
django-jalali-date==1.0.1
django-js-asset==2.0.0
django-rosetta==0.9.8
environs==9.5.0
filelock==3.8.0
idna==3.3
isodate==0.6.1
jdatetime==4.1.0
kavenegar==1.1.2
lxml==4.9.1
marshmallow==3.17.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
polib==1.1.1
py==1.11.0
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.4.0
pyparsing==3.0.9
python-decouple==3.6
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
urllib3==1.26.12
virtualenv==20.16.5
zeep==4.1.0
